Split concrete Contract class into an inteface and RatifiedContract
