+++ src/test/Lawyer.tools.test.ts
@@ -3,6 +3,7 @@ import assert from 'node:assert';
 import { OptionalType, RequiredType } from "../api/Types";
 import { Config as ContractConfig, Contract } from "../api/Contract";
 import { Lawyer } from "../api/Lawyer";
+import { createContract } from "../index";
 
 const LAWYER : Lawyer<Date> = new class implements Lawyer<Date> {
     createContract<X extends Date>(config?: ContractConfig<X> | undefined): RequiredType<Contract<X>>{
@@ -10,7 +11,7 @@ const LAWYER : Lawyer<Date> = new class implements Lawyer<Date> {
 
         copy.test ??= (instance: unknown): instance is X => instance === null || instance === undefined || instance instanceof Date;
 
-        return Contract.create<X>(copy);
+        return createContract<X>(copy);
     }
 
     isDeliverable<X extends Date>(instance: unknown): instance is OptionalType<X> {
